Added move thumbs by touch + fix for stuck thumbs
Now touching slider moves the closest thumb to touch point.
Also- fixed a bug- when two thumbs are at the extreme- trying to drag open the range doesn't work because the most extreme thumb gets the touch (instead of the inner thumb that needs to move). Also, dragging one thumb over the other opens the range in the opposite direction (the thumbs "flip sides" and continue to move).